Synchronization and Transfer Services

+++ datacollection/user_app/backend/app/services/box_service.py
@@ -1,3 +1,4 @@
+import concurrent
 import json
 import os
 import time
@@ -89,54 +90,71 @@ class BoxService:
 				self.client.folder(folder_id=parent_folder_id).create_subfolder(subfolder_name)).object_id
 		return subfolder_id
 	
-	def _upload_files_in_path(self, box_folder_id, folder_path):
+	def _upload_files_in_path(self, box_folder_id, folder_path, recording_id):
 		box_folder = self.client.folder(folder_id=box_folder_id)
 		box_files = {item for item in box_folder.get_items(item_type='file')}
 		for file_name in os.listdir(folder_path):
+			if not file_name.endswith(".zip") \
+					and not file_name.endswith(".pkl") \
+					and not file_name.endswith(".MP4") \
+					and not file_name.endswith(".mp4"):
+				logger.info(f"[{recording_id}] Skipping file: ", file_name)
+				continue
 			is_item_present = False
 			local_file_size = os.path.getsize(os.path.join(folder_path, file_name))
 			for box_item in box_files:
 				if file_name == box_item.name:
 					if local_file_size == box_item.size:
 						is_item_present = True
+						logger.info(f"[{recording_id}] File already present: ", file_name)
 					else:
+						logger.info(f"[{recording_id}] Deleting file as the file size does not match: ", file_name)
 						box_item.delete()
 						is_item_present = False
 			if not is_item_present:
 				file_path = os.path.join(folder_path, file_name)
+				logger.info(f"[{recording_id}] Uploading file: ", file_name)
 				box_folder.upload(file_path)
+				logger.info(f"[{recording_id}] Uploaded file: ", file_name)
 	
-	def _upload_folders_and_subfolders(self, parent_box_folder_id, parent_local_folder, folders):
+	def _upload_folders_and_subfolders(self, parent_box_folder_id, parent_local_folder, folders, recording_id):
 		for folder in folders:
 			box_folder_id = self._fetch_subfolder(parent_box_folder_id, folder)
 			local_folder_path = os.path.join(parent_local_folder, folder)
-			self._upload_files_in_path(box_folder_id, local_folder_path)
+			self._upload_files_in_path(box_folder_id, local_folder_path, recording_id)
 	
 	def upload_from_nas(self, recording, data_parent_directory):
 		activity_folder_id = self._fetch_activity_folder(self.activity_id_name_map[recording.activity_id])
 		recording_folder_id = self._fetch_subfolder(activity_folder_id, recording.id)
-		
+		recording_id = recording.id
 		data_recording_directory = os.path.join(data_parent_directory, recording.id)
 		raw_data_directory = os.path.join(data_recording_directory, const.RAW)
 		if os.path.exists(raw_data_directory):
+			logger.info(f"[{recording_id}] Uploading raw data")
 			raw_folder_id = self._fetch_subfolder(recording_folder_id, const.RAW)
 			self._upload_folders_and_subfolders(
 				raw_folder_id,
 				raw_data_directory,
-				[const.PV, const.DEPTH_AHAT, const.MICROPHONE, const.SPATIAL]
+				[const.PV, const.DEPTH_AHAT, const.MICROPHONE, const.SPATIAL],
+				recording_id
 			)
-		sync_data_directory = os.path.join(data_recording_directory, const.SYNCHRONIZED)
-		if os.path.exists(sync_data_directory):
-			sync_folder_id = self._fetch_subfolder(recording_folder_id, const.SYNCHRONIZED)
-			self._upload_folders_and_subfolders(
-				sync_folder_id,
-				sync_data_directory,
-				[const.PV, const.DEPTH_AHAT, const.SPATIAL]
-			)
+			logger.info(f"[{recording_id}] Raw data uploaded")
+		# sync_data_directory = os.path.join(data_recording_directory, const.SYNCHRONIZED)
+		# if os.path.exists(sync_data_directory):
+		# 	logger.info("Uploading synchronized data")
+		# 	sync_folder_id = self._fetch_subfolder(recording_folder_id, const.SYNCHRONIZED)
+		# 	self._upload_folders_and_subfolders(
+		# 		sync_folder_id,
+		# 		sync_data_directory,
+		# 		[const.PV, const.DEPTH_AHAT, const.SPATIAL]
+		# 	)
+		# 	logger.info("Synchronized data uploaded")
 		
+		logger.info(f"[{recording_id}] Uploading gopro data")
 		gopro_folder_id = self._fetch_subfolder(recording_folder_id, const.GOPRO)
 		local_gopro_path = os.path.join(data_recording_directory, const.GOPRO)
 		self._upload_files_in_path(gopro_folder_id, local_gopro_path)
+		logger.info(f"[{recording_id}] Gopro data uploaded")
 	
 	def upload_go_pro_360_video(self, recording, file_path):
 		logger.info(f'Uploading GoPro 360 video for recording {recording.id}')
@@ -189,6 +207,19 @@ class BoxService:
 		current_timestamp = time.strftime("%Y-%m-%d_%H-%M-%S")
 		timed_annotation_folder_id = self._fetch_subfolder(annotations_folder_id, f"{current_timestamp}")
 		self.client.folder(folder_id=timed_annotation_folder_id).upload(backup_annotation_file_path)
+	
+	def upload_in_threads(self, recordings, data_parent_directory, max_workers=5):
+		with concurrent.futures.ThreadPoolExecutor(max_workers=max_workers) as executor:
+			futures = {
+				executor.submit(self.upload_from_nas, recording, data_parent_directory): recording
+				for recording in recordings
+			}
+			for future in concurrent.futures.as_completed(futures):
+				recording = futures[future]
+				try:
+					future.result()
+				except Exception as exc:
+					logger.error('%r generated an exception: %s' % (recording, exc))
 
 
 if __name__ == '__main__':
